refactor: move timetables to /components/timetables
